Add VSCode workspace configuration
I wanted it to be a little bit easier to work with the playground within this repository and opted to implement this with a visual studio code workspace file. This file separates the playground folder in the files tree from the TIL entries, making it much easier to navigate to the playground folder as the today I learned category folders grow over time.

* Add VSCode workspace file
* Add VSCode settings file for root folder